+++ app/src/main/java/com/polant/projectsport/adapter/dialog/AdapterTodayFoodDialog.java
@@ -9,7 +9,7 @@ import android.widget.CursorAdapter;
 import android.widget.TextView;
 
 import com.polant.projectsport.R;
-import com.polant.projectsport.data.Database;
+import com.polant.projectsport.data.database.Database;
 
 /**
  * Created by Антон on 11.10.2015.
