+++ app/src/main/java/com/polant/projectsport/model/UserParametersInfo.java
@@ -1,4 +1,4 @@
-package com.polant.projectsport.data.model;
+package com.polant.projectsport.model;
 
 import android.content.Context;
